apply path compression to defmt logs
