+++ src/backtrace/pp.rs
@@ -1,8 +1,10 @@
 //! Pretty printing the backtrace
 
-use std::{borrow::Cow, path::Path};
+use std::borrow::Cow;
 
 use colored::Colorize as _;
+
+use crate::utils;
 
 use super::{symbolicate::Frame, Settings};
 
@@ -39,7 +41,7 @@ pub(crate) fn backtrace(frames: &[Frame], settings: &Settings) {
 
                 if let Some(location) = &subroutine.location {
                     let path = if settings.compress_cratesio_dep_paths {
-                        compress_cratesio_dep_path(&location.path)
+                        utils::compress_cratesio_dep_path(&location.path)
                     } else {
                         location.path.display().to_string()
                     };
@@ -70,60 +72,6 @@ pub(crate) fn backtrace(frames: &[Frame], settings: &Settings) {
                     break;
                 }
             }
-        }
-    }
-}
-
-// TODO use this for defmt logs
-fn compress_cratesio_dep_path(path: &Path) -> String {
-    if let Some(dep) = Dependency::from_path(path) {
-        format!("[{}]/{}", dep.name_version, dep.path.display())
-    } else {
-        path.display().to_string()
-    }
-}
-
-struct Dependency<'p> {
-    name_version: &'p str,
-    path: &'p Path,
-}
-
-impl<'p> Dependency<'p> {
-    // as of Rust 1.52.1 this path looks like this on Linux
-    // /home/some-user/.cargo/registry/src/github.com-0123456789abcdef/crate-name-0.1.2/src/lib.rs
-    // on Windows the `/home/some-user` part becomes something else
-    fn from_path(path: &'p Path) -> Option<Self> {
-        if !path.is_absolute() {
-            return None;
-        }
-
-        let mut components = path.components();
-        let _registry = components.find(|component| match component {
-            std::path::Component::Normal(component) => *component == "registry",
-            _ => false,
-        })?;
-
-        if let std::path::Component::Normal(src) = components.next()? {
-            if src != "src" {
-                return None;
-            }
-        }
-
-        if let std::path::Component::Normal(github) = components.next()? {
-            let github = github.to_str()?;
-            if !github.starts_with("github.com-") {
-                return None;
-            }
-        }
-
-        if let std::path::Component::Normal(name_version) = components.next()? {
-            let name_version = name_version.to_str()?;
-            Some(Dependency {
-                name_version,
-                path: components.as_path(),
-            })
-        } else {
-            None
         }
     }
 }
